Version 3.4.3
Clear the global thread table when an isolate is disposed (issue 1433).

Converted time zone name to UTF8 on Windows (issue 1290).

Limited the number of arguments in a function call to 32766 (issue 1413).

Compress sources of JS libraries in addition to the snapshot.

Fixed a bug in Lithium environment iteration.

Performance improvements on all platforms.

git-svn-id: http://v8.googlecode.com/svn/trunk@8218 ce2b1a6d-e550-0410-aec6-3dcde31c8c00